moving some dependencies around
